requirements.txt
fastmcp>=0.1.0
httpx>=0.24.0
python-dotenv>=1.0.0
cloudflare>=2.19.0